Splitting up most of neo_marker.rs
