+++ components/datetime/README.md
@@ -29,7 +29,7 @@ programmer to pick the calendar at compile time.
 ```rust
 use icu::calendar::{DateTime, Gregorian};
 use icu::datetime::neo::{DateTimeFormatter, FixedCalendarDateTimeFormatter};
-use icu::datetime::neo_marker::NeoYearMonthDayHourMinuteMarker;
+use icu::datetime::fieldset::NeoYearMonthDayHourMinuteMarker;
 use icu::datetime::neo_skeleton::NeoSkeletonLength;
 use icu::locale::{locale, Locale};
 use writeable::assert_try_writeable_eq;
