+++ components/datetime/src/input.rs
@@ -5,7 +5,7 @@
 //! A collection of utilities for representing and working with dates as an input to
 //! formatting operations.
 
-use crate::neo_marker::{DateInputMarkers, GetField, TimeMarkers, ZoneMarkers};
+use crate::scaffold::{DateInputMarkers, GetField, TimeMarkers, ZoneMarkers};
 use icu_calendar::any_calendar::AnyCalendarKind;
 use icu_calendar::{Date, Iso, Time};
 use icu_timezone::scaffold::IntoOption;
