+++ components/datetime/tests/datetime.rs
@@ -475,7 +475,7 @@ fn test_time_zone_format_configs() {
 
 #[test]
 fn test_time_zone_format_offset_seconds() {
-    use icu_datetime::{neo_marker::NeoTimeZoneOffsetMarker, neo_skeleton::NeoSkeletonLength};
+    use icu_datetime::{fieldset::NeoTimeZoneOffsetMarker, neo_skeleton::NeoSkeletonLength};
 
     let time_zone = TimeZoneInfo::from(UtcOffset::try_from_seconds(12).unwrap());
     let tzf = FixedCalendarDateTimeFormatter::<(), _>::try_new(
@@ -488,7 +488,7 @@ fn test_time_zone_format_offset_seconds() {
 
 #[test]
 fn test_time_zone_format_offset_not_set_debug_assert_panic() {
-    use icu_datetime::{neo_marker::NeoTimeZoneOffsetMarker, neo_skeleton::NeoSkeletonLength};
+    use icu_datetime::{fieldset::NeoTimeZoneOffsetMarker, neo_skeleton::NeoSkeletonLength};
 
     let time_zone = TimeZoneInfo {
         time_zone_id: TimeZoneIdMapper::new()
