+++ components/icu/examples/tui.rs
@@ -10,7 +10,7 @@ use icu::locale::locale;
 use icu::plurals::{PluralCategory, PluralRules};
 use icu::timezone::TimeZoneInfo;
 use icu_collections::codepointinvlist::CodePointInversionListBuilder;
-use icu_datetime::neo_marker::NeoYearMonthDayHourMinuteSecondTimeZoneOffsetMarker;
+use icu_datetime::fieldset::NeoYearMonthDayHourMinuteSecondTimeZoneOffsetMarker;
 use icu_datetime::neo_skeleton::NeoSkeletonLength;
 use icu_datetime::FixedCalendarDateTimeFormatter;
 use icu_timezone::CustomZonedDateTime;
